requirements.txt
Django==1.11.13
ruamel-yaml~=0.15
requests==2.19
emg-backlog-schema